Added README.md file.
This is after confusion with github using main and master as branches.
The master branch branch is what I pushed from the terminal and the main
was the default. I changed the default branch to master and deleted the
main branch. IDK if this is what I should have done.