Made fullassembly file Made IC chip version of ALU
Made a full assembly file for the control unit. I want to replace it
using the IC chip version of these because it makes it easier to copy
onto the real thing.

The IC chip version has a simpler version which includes a 4 bit adder
as an IC and the other one has the barebones, Logic gates. It was
painful.