Made ICs for PC, RAM Module and Reg
I finished making all the ICs I need for the entire computer. I finished
making them for the PC using an easy to use IC and same with the Reg.

The RAM Modules consists of the RAM, MDR and MAR registers too. These
are wired up and all work as expected.

NOTE: Some of the inputs are default high, make sure to remember which
are inverted.